Move safe_json_encode back to api.log as it's unused elsewhere
